Added standard input to the client, server responds to the test cases added
An interesting thing happens when an input is bigger than the buffer. Instead of doing something unexpected which I thought would happen, it just truncated was was bigger than the buffer. Not sure if I should ask the user to type again if the input exceeds the buffer size or to just ignore it.